+++ src/yamk/make.py
@@ -37,6 +37,7 @@ class Recipe:
         self._update_variables(globs, groups)
         self._update_requirements()
         self._update_commands()
+        self._add_implicit_variables()
 
     def _update_variables(self, globs, groups):
         self.vars = globs.add_batch(self.vars).add_batch([groups])
@@ -46,6 +47,10 @@ class Recipe:
 
     def _update_requirements(self):
         self.requires = lib.substitute_vars(self.requires, self.vars)
+
+    def _add_implicit_variables(self):
+        implicit_variables = {"target": self.target, "requirements": self.requires}
+        self.vars = self.vars.add_batch([implicit_variables])
 
 
 class MakeCommand: