+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add ${target} and ${requirements} as implicit variables
+
 ### Fixed
-- Fixed crashing bug when having lists or dicts as variable values
+- Fix crashing bug when having lists or dicts as variable values
 
 ## [0.5.0] - 2020-02-20
 ### Added