+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.6.0] - 2020-02-25
 ### Added
 - Add ${.target} and ${.requirements} as implicit variables
 - Allow aliases for targets